+++ INA228.cpp
@@ -1,8 +1,8 @@
 //    FILE: INA228.cpp
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.1.0
+// VERSION: 0.1.1
 //    DATE: 2024-05-09
-// PURPOSE: Arduino library for INA228 power sensor
+// PURPOSE: Arduino library for INA228 current and power sensor
 //     URL: https://github.com/RobTillaart/INA228
 //
 //  Read the datasheet for the details
@@ -317,7 +317,35 @@ uint16_t INA228::getShuntTemperatureCoefficent()
 //
 //  DIAGNOSE ALERT REGISTER 11
 //
+void INA228::setDiagnoseAlert(uint16_t flags)
+{
+  _writeRegister(INA228_DIAG_ALERT, flags);
+}
 
+void INA228::setDiagnoseAlertBit(uint8_t bit)
+{
+  uint16_t value = _readRegister(INA228_DIAG_ALERT, 2);
+  value |= (1 << bit);
+  _writeRegister(INA228_DIAG_ALERT, value);
+}
+
+void INA228::clrDiagnoseAlertBit(uint8_t bit)
+{
+  uint16_t value = _readRegister(INA228_DIAG_ALERT, 2);
+  value &= ~(1 << bit);
+  _writeRegister(INA228_DIAG_ALERT, value);
+}
+
+uint16_t INA228::getDiagnoseAlert()
+{
+  return _readRegister(INA228_DIAG_ALERT, 2);
+}
+
+uint16_t INA228::getDiagnoseAlertBit(uint8_t bit)
+{
+  uint16_t value = _readRegister(INA228_DIAG_ALERT, 2);
+  return (value >> bit) & 0x01;
+}
 
 
 ////////////////////////////////////////////////////////
@@ -325,8 +353,67 @@ uint16_t INA228::getShuntTemperatureCoefficent()
 //  THRESHOLD AND LIMIT REGISTERS 12-17
 //
 
+void INA228::setShuntOvervoltageTH(uint16_t threshold)
+{
+  _writeRegister(INA228_SOVL, threshold);
+}
 
+uint16_t INA228::getShuntOvervoltageTH()
+{
+  return _readRegister(INA228_SOVL, 2);
+}
 
+void INA228::setShuntUndervoltageTH(uint16_t threshold)
+{
+  _writeRegister(INA228_SUVL, threshold);
+}
+
+uint16_t INA228::getShuntUndervoltageTH()
+{
+  return _readRegister(INA228_SUVL, 2);
+}
+
+void INA228::setBusOvervoltageTH(uint16_t threshold)
+{
+  if (threshold > 0x7FFF) return;
+  _writeRegister(INA228_BOVL, threshold);
+}
+
+uint16_t INA228::getBusOvervoltageTH()
+{
+  return _readRegister(INA228_BOVL, 2);
+}
+
+void INA228::setBusUndervoltageTH(uint16_t threshold)
+{
+  if (threshold > 0x7FFF) return;
+  _writeRegister(INA228_BUVL, threshold);
+}
+
+uint16_t INA228::getBusUndervoltageTH()
+{
+  return _readRegister(INA228_BUVL, 2);
+}
+
+void INA228::setTemperatureOverLimitTH(uint16_t threshold)
+{
+  _writeRegister(INA228_TEMP_LIMIT, threshold);
+}
+
+uint16_t INA228::getTemperatureOverLimitTH()
+{
+  return _readRegister(INA228_TEMP_LIMIT, 2);
+}
+
+void INA228::setPowerOverLimitTH(uint16_t threshold)
+{
+  _writeRegister(INA228_POWER_LIMIT, threshold);
+}
+
+uint16_t INA228::getPowerOverLimitTH()
+{
+  return _readRegister(INA228_POWER_LIMIT, 2);
+}
 
 
 ////////////////////////////////////////////////////////